SunwayMRPi example

+++ include/Utils.hpp
@@ -178,7 +178,7 @@ long getCurrentTime()
 {
    struct timeval tv;
    gettimeofday(&tv,NULL);
-   return (long long)tv.tv_sec * 1000 + tv.tv_usec / 1000;
+   return tv.tv_sec * 1000 + tv.tv_usec / 1000;
 }
 
 int randomValue(int start, int end) {
